refactored project to work with Arduino ide
The arduino ide lets me input my data directly from the sketch contents and file name and compile the sketch beforehand.
This means that I deleleted all the classes that prompt the user for input and validates the responce.
I also separated tests and source files from eachother, which will make it easier to put the source files into the arduin ide, possibly with a jar file to encapsulate and hide them.

Potential bug:Some fields can be read to the arduino class generators as null or "null", I should allow both