Finishing touches interfacing with IDE
The new file structure allows these classes to be edited in eclpise and run with the main method, which calls a batch script.
The other valid option for enviornments would be to import the Arduino IDE as an eclipe project, but I didn't get that to work as well.

RunArduino.bat: created function to stop if error, copies files from this project into Arduino File Structure
EditorTab/SketchController,SketchController,SketchFile: Created dummy files so the project will compile in eclipse.